build: short circuit dev if any process fails
I was missing some required environment variables which was causing the
`next dev` process to fail, but the rest of the concurrent processes
started up with no problem and pushed the env var error up off the
screen.

This `--kill-others-on-fail` flag from the `concurrently` package tells
the whole concurrent dev process to fail-fast if any of them fail (have
a non-zero exit code).