feat: handle scenario where getTip is null
It's possible that the `getTip` function is directed to look up a Tip
that doesn't exist in Sanity. So the type needs to reflect that by
having a return type of `Tip | null`. The type errors then helpfully
pointed to some places where we needed to handle the `null` case.